Quotes_Django_deployed/project_quotes/pyproject.toml
[tool.poetry]
name = "hw13-django"
version = "0.1.0"
description = ""
authors = ["Denis Bilokon <greenjuiced@gmail.com>"]
readme = "README.md"
#packages = [{include = "hw13_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
scrapy = "^2.9.0"
beautifulsoup4 = "^4.12.2"
python-dotenv = "^1.0.0"
pydantic = "^1.10.7"
psycopg2-binary = "^2.9.6"
django-environ = "^0.10.0"
pymongo = "^4.3.3"
openai = "^0.27.7"



[tool.poetry.group.dev.dependencies]
openai = "^0.27.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
